General project structure redesign + middleware for error handling

+++ LevelApp.DAL/Repositories/User/UserRepository.cs
@@ -1,14 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
-using LevelApp.DAL.Models;
+using LevelApp.DAL.Entities;
 using LevelApp.DAL.Repositories.Base;
 using Microsoft.EntityFrameworkCore;
 
 namespace LevelApp.DAL.Repositories.User
 {
-    public class UserRepository: BaseRepository<CoreUser, long>, IUserRepository
+    public class UserRepository : BaseRepository<CoreUser, int>, IUserRepository
     {
         public UserRepository(DbContext context) : base(context)
         {
